add tests directory
